requirements.txt
# Core dependencies
python>=3.10
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
ultralytics>=8.1.0
opencv-python>=4.8.0
numpy>=1.25.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
albumentations>=1.3.1
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.66.0
PyYAML>=6.0
Pillow>=10.0.0

# Detectron2 (for Faster R-CNN / Mask R-CNN)
# Note: install from source if CUDA mismatch occurs
detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# Logging and reporting
tensorboard>=2.15.0
wandb>=0.17.0

# Visualization and plotting
plotly>=5.17.0
tabulate>=0.9.0

# Reproducibility and configuration
hydra-core>=1.3.2
omegaconf>=2.3.0

# Deployment and export
onnx>=1.15.0
onnxruntime>=1.16.0
torchsummary>=1.5.1

# Optional (Jetson / Raspberry Pi)
# install manually if running on ARM devices
# jetson-stats
# picamera2

# Testing and utilities
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0

# Documentation
markdown2>=2.4.10
